requirements.txt
datasets==4.0.0
matplotlib==3.10.6
numpy==2.3.3
peft==0.17.0
Pillow==11.3.0
Requests==2.32.5
scipy==1.16.2
soundfile==0.13.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.0.dev0
xformers==0.0.32.post2